+++ NEWS.md
@@ -1,4 +1,6 @@
 # arkhe 1.6.0.9000
+## New classes and methods
+* Add `scale_midpoint` to rescale a continuous vector to have specified minimum, midpoint and maximum.
 
 # arkhe 1.6.0
 ## New classes and methods